+++ styles/results.less
@@ -150,8 +150,6 @@
           margin-left: @padding-large-horizontal;
           margin-bottom: @padding-large-vertical;
           padding: @padding-base-vertical @padding-base-horizontal;
-
-          @import 'result-details';
         }
 
         .expanded-details {
